+++ js/mocks.js
@@ -1,0 +1,33 @@
+export const photoMockData = {
+  photoIDs: {min: 1, max: 25},
+  likesQuantity: {min: 25, max: 200},
+
+  photoDescriptions: [
+    'accusamus beatae ad facilis cum similique qui sunt',
+    'reprehenderit est deserunt velit ipsam',
+    'officia porro iure quia iusto qui ipsa ut modi',
+    'culpa odio esse rerum omnis laboriosam voluptate repudiandae',
+    'natus nisi omnis corporis facere molestiae rerum in',
+    'accusamus ea aliquid et amet sequi nemo',
+    'officia delectus consequatur vero aut veniam explicabo molestias',
+    'aut porro officiis laborum odit ea laudantium corporis'
+  ],
+};
+
+export const commentsMockData = {
+  commentIDs: {min: 1, max: 10000},
+  avatarIDs: {min: 1, max: 6},
+  messageQuantity: {min: 0, max: 30},
+  messageSentences: {min: 1, max: 2},
+
+  messages: [
+    'Всё отлично!',
+    'В целом всё неплохо. Но не всё.',
+    'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+    'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+    'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+    'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+  ],
+
+  names: ['Артём', 'Иван', 'Иннокентий', 'Бенджамин', 'Себастьян', 'Арнольд'],
+};
